% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melbourne}
\alias{melbourne}
\title{Melbourne}
\format{A data frame with 397 observations and 7 variables
\describe{
  \item{SA3_NAME16}{statistical area 3 name of the polygon}
  \item{GCC_NAME16}{}
  \item{GCC_CODE16}{}
  \item{SA2_MAIN16}{statistical area 2 code of the polygon}
  \item{AREASQKM16}{area of the SA2 polygon}
  \item{SA2_NAME16}{statistical area 2 name of the polygon}
  \item{SA4_CODE16}{statistical area 4 code of the polygon}
  \item{SA4_NAME16}{statistical area 4 name of the polygon}
  \item{geometry}{sfc geometry column}
}}
\usage{
melbourne
}
\description{
Polygons for Melbourne and the surrounding area
}
\details{
This data set is a subset of the Statistical Area Level 2 (SA2) ASGS
Edition 2016 data released by the Australian Bureau of Statistics
\url{ http://www.abs.gov.au }

The data is realsed under a Creative Commons Attribution 2.5 Australia licence
\url{https://creativecommons.org/licenses/by/2.5/au/}

The data has been down-cast from MULTIPOLYGONS to POLYGONS.
}
\keyword{datasets}
