% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_widget.R
\name{spatial_line}
\alias{spatial_line}
\title{Spatial line}
\usage{
spatial_line(sf, stroke_colour, stroke_opacity, stroke_width,
  legend = TRUE)
}
\arguments{
\item{sf}{\code{sf} object}

\item{stroke_colour}{string specifying column of \code{sf} to use for the stroke colour,
or a single value to apply to all rows of data}

\item{stroke_opacity}{string specifying column of \code{sf} to use for the stroke opacity,
or a single value to apply to all rows of data}

\item{stroke_width}{string specifying column of \code{sf} to use for the stroke width,
or a single value to apply to all rows of data}

\item{legend}{logical indicating if legend data will be returned}
}
\description{
Spatial line
}
