% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widget_arcs}
\alias{widget_arcs}
\title{Origin Destination points between Sydney, Australia and other capitals cities}
\format{
A sf object with 199 observations and 6 variables
\describe{
  \item{country_from}{origin country}
  \item{capital_from}{origin capital}
  \item{country_to}{destination country}
  \item{capital_to}{destination capital}
  \item{origin}{sfc geometry column}
  \item{destination}{sfc geometry column}
}
}
\usage{
widget_arcs
}
\description{
A simple feature \code{sf} object with two \code{sfc} columns, "origin" and "destination"
}
\keyword{datasets}
