% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widget_capitals}
\alias{widget_capitals}
\title{Capital cities for each country}
\format{A sf object with 200 observations and 4 variables
\describe{
  \item{country}{country name}
  \item{capital}{capital name}
  \item{geometry}{sfc geometry column}
}}
\usage{
widget_capitals
}
\description{
A simple feature \code{sf} object containing the coordinates of 200 capitical cities in the world
}
\keyword{datasets}
