% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_widget.R
\name{widget_line}
\alias{widget_line}
\title{Widget Line}
\usage{
widget_line(data, stroke_colour, stroke_opacity, stroke_width,
  legend = TRUE)
}
\arguments{
\item{stroke_colour}{string specifying column of \code{sf} to use for the stroke colour,
or a single value to apply to all rows of data}

\item{stroke_opacity}{string specifying column of \code{sf} to use for the stroke opacity,
or a single value to apply to all rows of data}

\item{stroke_width}{string specifying column of \code{sf} to use for the stroke width,
or a single value to apply to all rows of data}

\item{legend}{logical indicating if legend data will be returned}

\item{sf}{\code{sf} object}
}
\description{
Widget Line
}
\examples{

## use default stroke options
l <- widget_line( roads, legend = F )

}
