% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_widget.R
\name{widget_point}
\alias{widget_point}
\title{Widget Point}
\usage{
widget_point(data, fill_colour, fill_opacity, lon = NULL, lat = NULL,
  legend = TRUE)
}
\arguments{
\item{fill_opacity}{}

\item{lat}{}

\item{legend}{logical indicating if legend data will be returned}
}
\description{
Widget Point
}
