% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_widget.R
\name{widget_polygon}
\alias{widget_polygon}
\title{Widget Polygon}
\usage{
widget_polygon(data, stroke_colour, stroke_opacity, stroke_width,
  fill_colour, fill_opacity, legend = TRUE)
}
\arguments{
\item{stroke_colour}{string specifying column of \code{sf} to use for the stroke colour,
or a single value to apply to all rows of data}

\item{stroke_opacity}{string specifying column of \code{sf} to use for the stroke opacity,
or a single value to apply to all rows of data}

\item{stroke_width}{string specifying column of \code{sf} to use for the stroke width,
or a single value to apply to all rows of data}

\item{fill_opacity}{}

\item{legend}{logical indicating if legend data will be returned}
}
\description{
Widget Polygon
}
\examples{

l <- widget_polygon( melbourne, legend = F)
l <- widget_polygon( melbourne, fill_colour = "AREASQKM16", legend = F)

}
