% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_widget.R
\name{widget_line}
\alias{widget_line}
\title{Widget Line}
\usage{
widget_line(data, stroke_colour = NULL, stroke_opacity = NULL,
  stroke_width = NULL, legend = TRUE, json_legend = TRUE)
}
\arguments{
\item{data}{\code{sf} object}

\item{stroke_colour}{string specifying column of \code{sf} to use for the stroke colour,
or a single value to apply to all rows of data}

\item{stroke_opacity}{string specifying column of \code{sf} to use for the stroke opacity,
or a single value to apply to all rows of data}

\item{stroke_width}{string specifying column of \code{sf} to use for the stroke width,
or a single value to apply to all rows of data}

\item{legend}{logical indicating if legend data will be returned}

\item{json_legend}{logical indicating if the legend will be returned as json}
}
\description{
Converts an `sf` object with LINESTRING geometriers into JSON for plotting in an htmlwidget
}
\examples{

## use default stroke options
l <- widget_line( widget_roads, legend = TRUE )

}
